DAOS-2113 control: Address review feedback
Move TestMain() and other common items to pbin_test.go
